Close #3.

+++ ManagedSpyLib/Mem.cpp
@@ -107,7 +107,7 @@ bool MemoryStore::StoreData(CAtlFileMapping<SharedData>& memory, Object^ data, L
 	if (((SharedData*)memory) == NULL) {
 		return false;
 	}
-	((SharedData*)memory)->Size = (SIZE_T)stream->Length;
+	((SharedData*)memory)->Size = static_cast<std::uint32_t>(stream->Length);
 	BYTE* pdata = (BYTE*)((SharedData*)memory)->Data;
 
 	array<unsigned char>^ streamdata = stream->GetBuffer();
